requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
openai>=1.3.0
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
yookassa>=3.0.0
